Ändrade tillbaka tagId till Int32

+++ Handler/PLC/Logging/SqlClasses/LastValue.cs
@@ -3,7 +3,7 @@ namespace OptimaValue
 {
     public class LastValue
     {
-        public decimal tag_id { get; set; }
+        public int tag_id { get; set; }
         public object value { get; set; }
         public DateTime logDate { get; set; }
     }
